# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Book Management System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get list of authors",
                "description": "Returns list of authors",
                "operationId": "b1de4facd6962392c6fde61f8a7b097a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/authorsearch": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Search authors",
                "description": "Search authors by name",
                "operationId": "e433c16fdc852ddf0adb6cfa82645425",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query string",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Author"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/author": {
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create a new author",
                "description": "Creates a new author with the given data",
                "operationId": "5d7a5ee4c26df8bab9e8bed5f32ac0d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "biography",
                                    "book_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "biography": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/author/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get an author by ID",
                "description": "Returns a single author by ID",
                "operationId": "f5da13c0fe7f7b4671318c86ce983ab9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/authorbooks/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get the books written by each author",
                "description": "Get the books written by each author",
                "operationId": "3ae26d2a645c223c88ea6e25df3ac76a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Author"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update an author",
                "description": "Update an author by ID",
                "operationId": "8797e8c11aa3838017483c14703f088c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "biography",
                                    "book_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "biography": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auteur modifié avec succès"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete an author",
                "description": "Deletes an author by ID",
                "operationId": "c56695bdfa8aa06b4a00f609c64bdc40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Auteur supprimé avec succès"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Author": {
                "title": "Author",
                "required": [
                    "name",
                    "biography",
                    "book_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the author",
                        "type": "string"
                    },
                    "biography": {
                        "description": "Biography of the author",
                        "type": "string"
                    },
                    "book_id": {
                        "description": "ID of the book",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}